fixbug: Can not execute native `pwsh` command
For more details, the reason why we could not execute a list of native
`pwsh` command while running the `go` function `run()` by using the `os`
package is still a big question that has been needed to resolve (?).
Some cases still working quite normally like all the `git` commands, my
prediction is all about the restrictions that the creator of `cmd`
package had set to it. Or maybe just about the boundary of subprocess in
Windows only (?).